Split fragment color rendering helpers
