+++ crates/pixel_world_bridge/src/render.rs
@@ -3,6 +3,10 @@ use std::collections::{HashMap, HashSet};
 use super::*;
 use bevy::ecs::system::SystemParam;
 
+#[path = "render_fragment_visuals.rs"]
+mod fragment_visuals;
+use fragment_visuals::{fragment_color, fragment_inset_color, fragment_shadow_color};
+
 const LOCATION_HIT_HALF_SIZE: f64 = 8.0;
 const AGENT_HIT_HALF_SIZE: f64 = 8.0;
 const FRAGMENT_HIDDEN_THRESHOLD_PX: f64 = 1.5;
@@ -422,36 +426,6 @@ fn grid_geometry(layout: &GridLayoutKey) -> (f64, f64, f64, f64, Color) {
         layout.offset_y_milli as f64 / 1000.0,
         layout.width as f64,
         Color::srgba_u8(99, 179, 255, 26),
-    )
-}
-
-fn fragment_color(fragment: &FragmentTerrainPatch, lod: FragmentTerrainLod) -> Color {
-    let alpha = fragment_alpha(fragment, lod);
-    Color::srgba_u8(
-        fragment.color[0],
-        fragment.color[1],
-        fragment.color[2],
-        (alpha.clamp(0.0, 1.0) * 255.0).round() as u8,
-    )
-}
-
-fn fragment_inset_color(fragment: &FragmentTerrainPatch) -> Color {
-    let alpha = fragment_alpha(fragment, FragmentTerrainLod::Detail);
-    Color::srgba_u8(
-        fragment.color[0].saturating_mul(3) / 5,
-        fragment.color[1].saturating_mul(3) / 5,
-        fragment.color[2].saturating_mul(3) / 5,
-        (alpha.clamp(0.0, 1.0) * 255.0).round() as u8,
-    )
-}
-
-fn fragment_shadow_color(fragment: &FragmentTerrainPatch, lod: FragmentTerrainLod) -> Color {
-    let alpha = fragment_alpha(fragment, lod) * f64::from(FRAGMENT_SHADOW_ALPHA_CAP);
-    Color::srgba_u8(
-        fragment.color[0].saturating_mul(2) / 5,
-        fragment.color[1].saturating_mul(2) / 5,
-        fragment.color[2].saturating_mul(2) / 5,
-        (alpha.clamp(0.0, 1.0) * 255.0).round() as u8,
     )
 }
